Changed to use heex.
- LogicModulesGenerator renamed to ModulesGenerator.
- Modules generation splitted into separated modules.
